w14-7 data type fixed

+++ python_apps/w14_hw_read_df_from_s3_write_postgres.py
@@ -70,10 +70,10 @@ df['DISCOUNT'] = df['DISCOUNT'].astype(float)
 df['SP'] = df['SP'].astype(float)
 df['DATE'] = pd.to_datetime(df['Date'], format='%Y-%m-%d')  # Assuming the date format is year-month-day
 
-# Extract day, month, and year from 'Date' column
-df['Day'] = df['Date'].dt.day
-df['Month'] = df['Date'].dt.month
-df['Year'] = df['Date'].dt.year
+# Extract day, month, and year from 'Date' column and cast to integer
+df['Day'] = df['Date'].dt.day.astype(int)
+df['Month'] = df['Date'].dt.month.astype(int)
+df['Year'] = df['Date'].dt.year.astype(int)
 
 # 6. Drop original 'Date' column
 df = df.drop(columns=['Date'])
